改善章节选择底栏的操作体验
